pyproject.toml
[tool.poetry]
name = "smse-backend"
version = "0.1.0"
description = ""
authors = ["Ahmedsaed <ahmedsaed2652003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.7.0"
python-dotenv = "^1.0.1"
pgvector = "^0.3.6"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# add entry point
[tool.poetry.scripts]
smse-backend = "smse_backend.app:main"